requirements.txt
asgiref==3.8.1
bleach==6.2.0
boto3==1.38.46
botocore==1.38.46
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
crispy-bootstrap5==2025.6
cryptography==45.0.4
dj-database-url==3.0.0
Django==5.2.1
django-allauth==65.9.0
django-ckeditor==6.7.3
django-ckeditor-5==0.2.18
django-crispy-forms==2.4
django-debug-toolbar==5.2.0
django-distill==3.2.7
django-environ==0.12.0
django-jazzmin==3.0.1
django-js-asset==3.1.2
django-storages==1.14.6
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
requests==2.32.3
s3transfer==0.13.0
six==1.17.0
sqlparse==0.5.3
stripe==12.2.0
typing_extensions==4.14.0
urllib3==2.4.0
webencodings==0.5.1
asgiref==3.8.1
bleach==6.2.0
boto3==1.38.46
botocore==1.38.46
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
crispy-bootstrap5==2025.6
cryptography==45.0.4
dj-database-url==3.0.0
Django==5.2.1
django-allauth==65.9.0
django-ckeditor==6.7.3
django-ckeditor-5==0.2.18
django-crispy-forms==2.4
django-debug-toolbar==5.2.0
django-distill==3.2.7
django-environ==0.12.0
django-jazzmin==3.0.1
django-js-asset==3.1.2
django-storages==1.14.6
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
requests==2.32.3
s3transfer==0.13.0
six==1.17.0
sqlparse==0.5.3
stripe==12.2.0
typing_extensions==4.14.0
urllib3==2.4.0
webencodings==0.5.1
